Restructure bittide-instances
The new file structure is as follows:
bittide-instances
└─ src
   └─ Bittide
      └─ Instances
         ├─ Pnr
         └─ Hitl
