+++ bittide-instances/bittide-instances.cabal
@@ -97,28 +97,31 @@ library
   import: common-options
   hs-source-dirs: src
   exposed-modules:
-    Bittide.Instances.Calendar
-    Bittide.Instances.ClockControl
-    Bittide.Instances.Counter
     Bittide.Instances.Domains
-    Bittide.Instances.ElasticBuffer
     Bittide.Instances.Hacks
-    Bittide.Instances.MVPs
-    Bittide.Instances.ProcessingElement
-    Bittide.Instances.ScatterGather
-    Bittide.Instances.Si539xSpi
-    Bittide.Instances.StabilityChecker
-    Bittide.Instances.Synchronizer
 
     -- Hardware-in-the-loop tests
-    Bittide.Instances.Tests.BoardTest
-    Bittide.Instances.Tests.FincFdec
-    Bittide.Instances.Tests.FullMeshHwCc
-    Bittide.Instances.Tests.FullMeshSwCc
-    Bittide.Instances.Tests.FullMeshHwCc.IlaPlot
-    Bittide.Instances.Tests.SyncInSyncOut
-    Bittide.Instances.Tests.Transceivers
-    Bittide.Instances.Tests.VexRiscv
+    Bittide.Instances.Hitl.BoardTest
+    Bittide.Instances.Hitl.FincFdec
+    Bittide.Instances.Hitl.FullMeshHwCc
+    Bittide.Instances.Hitl.FullMeshSwCc
+    Bittide.Instances.Hitl.FullMeshHwCc.IlaPlot
+    Bittide.Instances.Hitl.SyncInSyncOut
+    Bittide.Instances.Hitl.Transceivers
+    Bittide.Instances.Hitl.VexRiscv
+
+    -- Place and route tests
+    Bittide.Instances.Pnr.Calendar
+    Bittide.Instances.Pnr.ClockControl
+    Bittide.Instances.Pnr.Counter
+    Bittide.Instances.Pnr.ElasticBuffer
+    Bittide.Instances.Pnr.MVPs
+    Bittide.Instances.Pnr.ProcessingElement
+    Bittide.Instances.Pnr.ScatterGather
+    Bittide.Instances.Pnr.Si539xSpi
+    Bittide.Instances.Pnr.StabilityChecker
+    Bittide.Instances.Pnr.Synchronizer
+
     Project.FilePath
