+++ bittide-instances/src/Bittide/Instances/Hitl/FullMeshHwCc.hs
@@ -22,7 +22,7 @@
 -- This test will succeed if all clocks have been stable for 5 seconds. Note:
 -- this doesn't test reframing yet.
 --
-module Bittide.Instances.Tests.FullMeshSwCc where
+module Bittide.Instances.Hitl.FullMeshHwCc where
 
 import Clash.Prelude (withClockResetEnable)
 import Clash.Explicit.Prelude
@@ -55,8 +55,8 @@ import Bittide.ProcessingElement.Util (memBlobsFromElf)
 import Bittide.SharedTypes (Bytes, ByteOrder(BigEndian))
 import Bittide.Transceiver
 
-import Bittide.Instances.Tests.FullMeshHwCc.IlaPlot
-import Bittide.Instances.MVPs (stickyBits, speedChangeToPins, FINC, FDEC)
+import Bittide.Instances.Hitl.FullMeshHwCc.IlaPlot
+import Bittide.Instances.Pnr.MVPs (stickyBits, speedChangeToPins, FINC, FDEC)
 
 import Clash.Class.Counter
 import Clash.Cores.Xilinx.GTH
@@ -101,7 +101,7 @@ unitCS = CSignal (pure ())
 -- | Instantiates a RiscV core that copies instructions coming from a hardware
 -- implementation of Callisto (see 'fullMeshHwTest') and copies it to a register
 -- tied to FINC/FDEC.
-fullMeshRiscvTest ::
+fullMeshRiscvCopyTest ::
   forall dom .
   KnownDomain dom =>
   Clock dom ->
@@ -112,7 +112,7 @@ fullMeshRiscvTest ::
   ( "FINC" ::: Signal dom Bool
   , "FDEC" ::: Signal dom Bool
   )
-fullMeshRiscvTest clk rst callistoResult dataCounts = unbundle fIncDec
+fullMeshRiscvCopyTest clk rst callistoResult dataCounts = unbundle fIncDec
  where
   (_, CSignal fIncDec) = toSignals
     ( circuit $ \unit -> do
@@ -179,7 +179,7 @@ fullMeshRiscvTest clk rst callistoResult dataCounts = unbundle fIncDec
 
     root <- runIO findProjectRoot
 
-    let elfPath = root </> "_build/cargo/firmware-binaries/riscv32imc-unknown-none-elf/release/clock-control"
+    let elfPath = root </> "_build/cargo/firmware-binaries/riscv32imc-unknown-none-elf/release/clock-control-reg-cpy"
 
     memBlobsFromElf BigEndian elfPath Nothing)
 
@@ -191,7 +191,7 @@ fullMeshRiscvTest clk rst callistoResult dataCounts = unbundle fIncDec
   -}
   peConfig =
     PeConfig
-      (0b10 :> 0b01 :> 0b11 :> 0b00 :> Nil)
+      (0b10 :> 0b01 :> 0b00 :> 0b11 :> Nil)
       (Reloadable $ Blob iMem)
       (Reloadable $ Blob dMem)
 
@@ -395,7 +395,7 @@ trueFor5s clk rst =
   goOutput = (== maxBound)
 
 -- | Top entity for this test. See module documentation for more information.
-fullMeshSwCcTest ::
+fullMeshHwCcTest ::
   "SMA_MGT_REFCLK_C" ::: DiffClock Basic200 ->
   "SYSCLK_300" ::: DiffClock Basic300 ->
   "SYNC_IN" ::: Signal Basic125 Bool ->
@@ -416,7 +416,7 @@ fullMeshSwCcTest ::
       , "CSB"       ::: Signal Basic125 Bool
       )
   )
-fullMeshSwCcTest refClkDiff sysClkDiff syncIn rxns rxps miso =
+fullMeshHwCcTest refClkDiff sysClkDiff syncIn rxns rxps miso =
   (txns, txps, (riscvFinc, riscvFdec), syncOut, spiDone, spiOut)
  where
   refClk = ibufds_gte3 refClkDiff :: Clock Basic200
@@ -424,61 +424,67 @@ fullMeshSwCcTest refClkDiff sysClkDiff syncIn rxns rxps miso =
   (sysClk, sysLock0) = clockWizardDifferential (SSymbol @"SysClk") sysClkDiff noReset
   sysLock1 = xpmCdcSingle sysClk sysClk sysLock0 -- improvised reset syncer
   sysRst = unsafeFromActiveLow sysLock1
-
-  syncIn1 = (startTest .&&.)
-    $ unsafeToActiveLow
-    $ resetGlitchFilter (SNat @1024) sysClk
-    $ unsafeFromActiveLow
-    $ xpmCdcSingle sysClk sysClk syncIn
-
-  syncInChangepoints =
-    changepoints sysClk syncStartRst enableGen syncIn1
-
-  syncInRst =
-      unsafeFromActiveLow
-    $ sticky sysClk sysRst
-    $ isRising sysClk sysRst enableGen False
-      syncIn1
-
-  testRst = sysRst `orReset` syncInRst
-
-  syncStart =
-      sticky sysClk sysRst
-    $ isFalling sysClk sysRst enableGen False
-      syncIn1
-
-  syncStartRst = unsafeFromActiveLow syncStart
-
-  startBeforeAllUp = sticky sysClk sysRst
-    (syncStart .&&. ((not <$> allUp) .||. transceiversFailedAfterUp))
-
+    `orReset` unsafeFromActiveLow startTest
+  --
+  -- 'syncOutGenerator' is used to drive the 'SYNC_OUT' signal, which
+  -- is only connected for the last node in the network and wired back
+  -- to 'SYNC_IN' of all nodes from there.
+  --
+  -- Note that all nodes are in reset before their local 'startTest' VIO
+  -- signal gets asserted, as 'startTest' is directly driving 'sysRst'.
+  -- Thus, for the other nodes to capture the 'SYNC_OUT' signal correctly,
+  -- the node receiving the `startTest` rising edge last must be the one
+  -- with it's 'SYNC_OUT' physically connected to the 'SYNC_IN' of all
+  -- nodes in the network. This assumption is tested by
+  -- 'Bittide.Instances.Hitl.SyncInSyncOut'.
   syncOut =
       dflipflop sysClk
     $ syncOutGenerator sysClk startTest
     $ trueFor5s sysClk testRst allUp
 
-  --  syncOutGenerator sysClk sysRst enableGen
-  --    (startTest, trueFor5s sysClk sysRst allUp)
+  -- first synchronize SYNC_IN to the local clock and filter from
+  -- potential glitches
+  syncIn1 =
+      unsafeToActiveLow
+    $ resetGlitchFilter (SNat @1024) sysClk
+    $ unsafeFromActiveLow
+    $ xpmCdcSingle sysClk sysClk syncIn
 
+  -- generate a pulse on every change of SYNC_IN
+  syncInChangepoints =
+    changepoints sysClk sysRst enableGen syncIn1
+
+  -- recover the activity and readiness states from SYNC_IN
+  (syncActive, syncStart) = unbundle $ syncInRecover sysClk sysRst syncIn1
+
+  -- tests are reset with on `sysRst` or if not synchronously active
+  testRst = sysRst `orReset` unsafeFromActiveLow syncActive
+
+  -- checks that tests are not synchronously start before all
+  -- transceivers are up
+  startBeforeAllUp = sticky sysClk testRst
+    (syncStart .&&. ((not <$> allUp) .||. transceiversFailedAfterUp))
+
+  -- generate the global timestamp from the synchronous rising and
+  -- falling edges of SYNC_IN
   globalTimestamp :: Signal Basic125 GlobalTimestamp
-  globalTimestamp = register sysClk syncStartRst enableGen (0,0) $
+  globalTimestamp = register sysClk testRst enableGen (0,0) $
     mux syncInChangepoints
       (((+1) *** const 0) <$> globalTimestamp)
       (second (+1) <$> globalTimestamp)
 
   -- calibrate over the first 200 sync pulses
   calibrate =
-    moore sysClk syncStartRst enableGen
+    moore sysClk testRst enableGen
       (\s -> bool s $ satSucc SatBound s)
       (/= maxBound)
       (minBound :: Index 200)
       syncInChangepoints
 
   calibrationDone =
-    isFalling sysClk sysRst enableGen False calibrate
+    isFalling sysClk testRst enableGen False calibrate
 
-  -- TODO: not implemented yet
-  syncEnd = pure False
+  syncEnd = isFalling sysClk testRst enableGen False syncActive
 
   (   txns, txps, _hwFincFdecs, callistoClock, callistoResult, callistoReset
     , dataCounts, stats, spiDone, spiOut, transceiversFailedAfterUp, allUp
@@ -486,7 +492,7 @@ fullMeshSwCcTest refClkDiff sysClkDiff syncIn rxns rxps miso =
     fullMeshHwTest refClk sysClk testRst IlaControl{..} rxns rxps miso
 
   (riscvFinc, riscvFdec) =
-    fullMeshRiscvTest callistoClock callistoReset callistoResult dataCounts
+    fullMeshRiscvCopyTest callistoClock callistoReset callistoResult dataCounts
 
   stats0 :> stats1 :> stats2 :> stats3 :> stats4 :> stats5 :> stats6 :> Nil = stats
 
@@ -572,8 +578,8 @@ fullMeshSwCcTest refClkDiff sysClkDiff syncIn rxns rxps miso =
       (failAfterUps  <$> stats6)
 -- XXX: We use an explicit top entity annotation here, as 'makeTopEntity'
 --      generates warnings in combination with 'Vec'.
-{-# ANN fullMeshSwCcTest Synthesize
-  { t_name = "fullMeshSwCcTest"
+{-# ANN fullMeshHwCcTest Synthesize
+  { t_name = "fullMeshHwCcTest"
   , t_inputs =
     [ (PortProduct "SMA_MGT_REFCLK_C") [PortName "p", PortName "n"]
     , (PortProduct "SYSCLK_300") [PortName "p", PortName "n"]
