+++ bittide-instances/src/Bittide/Instances/Hitl/FullMeshHwCc/IlaPlot.hs
@@ -6,7 +6,7 @@
 {-# LANGUAGE FlexibleContexts #-}
 
 {-# OPTIONS_GHC -fconstraint-solver-iterations=20 #-}
-module Bittide.Instances.Tests.FullMeshHwCc.IlaPlot
+module Bittide.Instances.Hitl.FullMeshHwCc.IlaPlot
   ( PlotData(..)
   , IlaControl(..)
   , RfStageChange(..)
